requirements.txt
pycryptodome
web3 
